# Txhandler: Improve dedupe hit rate