@iansuvak 2023-01-18
It's about 2.5-3% of the transactions that we see are duplicates and could reasonably only check the cur to avoid having to rehash with the old salt. 

Benchmarks for that are here which looks like 4% slowdown on the thread=16 case.

The real cost to weigh is the cost of this slowdown vs the cost of re-processing these through the TxHandler right?
Either way happy to close if consensus is we don't want this

:
```
BenchmarkDigestCaches/data.digestCacheMaker/threads=1-10         	 1558552	       787.0 ns/op	     206 B/op	       2 allocs/op
BenchmarkDigestCaches/data.saltedCacheMaker/threads=1-10         	 1000000	      1076 ns/op	     220 B/op	       4 allocs/op
BenchmarkDigestCaches/data.digestCacheMaker/threads=4-10         	 2173646	       569.7 ns/op	     164 B/op	       2 allocs/op
BenchmarkDigestCaches/data.saltedCacheMaker/threads=4-10         	  958526	      1368 ns/op	     225 B/op	       4 allocs/op
BenchmarkDigestCaches/data.digestCacheMaker/threads=16-10        	 1612461	       794.1 ns/op	     201 B/op	       2 allocs/op
BenchmarkDigestCaches/data.saltedCacheMaker/threads=16-10        	  956253	      1371 ns/op	     225 B/op	       4 allocs/op
BenchmarkDigestCaches/data.digestCacheMaker/threads=128-10       	 1465798	       820.3 ns/op	     216 B/op	       2 allocs/op
BenchmarkDigestCaches/data.saltedCacheMaker/threads=128-10       	  942253	      1399 ns/op	     227 B/op	       4 allocs/op
```