@iansuvak 2023-01-18
> That's right, the cost of the slowdown for all cache checks vs the reprocessing cost of the ones that got through. Do you know how many total transactions were processed during the period you were looking at? Not just the ones that made it through the filter but the raw TX message count that hit this code — that would push the duplicate rate you're seeing of 2.5%-3% a lot further down I assume?

The TX messages that didn't make it through the filter by definition wouldn't hit this code no? They got filtered out using RLocks only right? The raw number is a lot larger of course but it still seems to me that this would be worth it iff:

`2.5% * t<sub>handleTxn</sub> >= ~120ns`
`t<sub>handleTxn</sub> >= ~4800ns`

Do we know the average or worst case t above?

> (Also these changes are only affecting saltedCacheMaker not the digestCacheMaker so the benchmarks are only relevant for half the results you're posting)

Yup, Just posted the whole output of the single Benchmark function without editing it.